tests/wpt/web-platform-tests/eventsource/format-newlines.htm

<!doctype html>
<html>
  <head>
    <title>EventSource: newline fest</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var source = new EventSource("resources/message.py?message=data%3Atest%0D%0Adata%0Adata%3Atest%0D%0A%0D&newline=none")
        source.onmessage = function(e) {
          test.step(function() {
            assert_equals("test\n\ntest", e.data)
            source.close()
          })
          test.done()
        }
      })
    </script>
  </body>
</html>
